backend/pyproject.toml
[project]
name = "pokemon-champions-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "python-multipart==0.0.9",
    "pydantic>=2.0",
    "opencv-python-headless==4.10.0.84",
    "numpy==1.26.4",
    "requests==2.32.3",
    "beautifulsoup4==4.12.3",
    "anthropic>=0.40.0",
    "litellm>=1.0.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.2",
    "pytest-asyncio==0.23.8",
    "httpx==0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
